Refactored RemoteKeys to use PartitionRange and Scripting.Dictionary.
No longer runs Sort on the Worksheet everytime we
call GetIDsFromQuery(). On first run it will run
PartitionRange and run a Sort. Subsequent runs will
used the cached Keys in a Dictionary. Newly created
Keys will be added to this Dictionary without needing
to rerun PartitionRange and Worksheet Sort.

Appends new Keys to end of table instead of clearing
entire Worksheet and replacing it with every Key from
memory.

Fixed bug when creating new
KeyPaths and there is only one row which returns
a single Variant instead of a Variant Array.