View History (Timelime) updates on selection change.
The UI for viewing the history of a cell now watches for
SelectionChange events on the Worksheet. When the selection
changes, it will try and rerun the ValueTimelineQuery.
If it succeeds, it updates the UserForm. If not, it clears
out the values in the UserForm (e.g., cell is in a column
that is not mapped, or cell is in a Key Column)